Split index.js to enforce separation of concerns
